Fixed the optimclass test
The stats::fisher.test() function can not calculate the test statistic when the species-cluster contingency table is 1x2 which happens when all observations of a species is found in all clusters. That is, there are no instances when a species is not found in any of the clusters. The work around solution is checking the dimensions of the contingency table first and returning a not significant p-value if the error would show up.